Fix 2 bugs so the password feature actually works :)
Two bugs:
1) The part that reads the password-string from our secrets file, adds
a \n. So if our password is "abc", the password was actually read as
"abc
", including the newline. This was never noticed because nobody ever
tried with a different password and bug nr 2)

2) It seems that the application.yaml file is read AFTER we read the
password from a text file. So by defining the password in the
application.yaml file, that password was always used! Their might be a
way to fix this, but my Java Spring Boot knowledge is not good enough.